# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forma-te API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "moises-alberto@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a paginated list of courses",
                "description": "Returns a paginated list of courses based on the provided query parameters.",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Additional filter for searching courses",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Course"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/course/{courseId}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get course by ID",
                "description": "Returns details of a single course based on the provided course ID",
                "operationId": "getCourseById",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "ID of course to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update a course",
                "description": "Update an existing course by ID",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the course to update.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUpdateCourseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseStoreResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate entry for the course URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Duplicate entry for the course URL"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred while updating the course",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred while updating the course"
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "Error details here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course",
                "description": "Deletes an existing course based on the provided ID.",
                "operationId": "destroyCourse",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "ID of the course to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course deleted successfully"
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreUpdateCourseRequest": {
                "required": [
                    "category_id",
                    "name"
                ],
                "properties": {
                    "category_id": {
                        "description": "The ID of the category",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the course",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 5
                    },
                    "description": {
                        "description": "The description of the course",
                        "type": "string",
                        "nullable": true
                    },
                    "short_name": {
                        "description": "The short name of the course",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "url": {
                        "description": "The URL of the course",
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "description": "The image of the course",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "file": {
                        "description": "The file associated with the course",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "type": {
                        "description": "The type of the course",
                        "type": "string",
                        "nullable": true
                    },
                    "code": {
                        "description": "The unique code of the course",
                        "type": "string",
                        "nullable": true
                    },
                    "total_hours": {
                        "description": "The total hours of the course",
                        "type": "string",
                        "nullable": true
                    },
                    "published": {
                        "description": "The published status of the course",
                        "type": "boolean",
                        "nullable": true
                    },
                    "free": {
                        "description": "The free status of the course",
                        "type": "boolean",
                        "nullable": true
                    },
                    "price": {
                        "description": "The price of the course",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CourseStoreResource": {
                "properties": {
                    "id": {
                        "description": "The ID of the course",
                        "type": "integer"
                    },
                    "category_id": {
                        "description": "The ID of the category",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "The ID of the user",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the course",
                        "type": "string"
                    },
                    "url": {
                        "description": "The URL of the course",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the course",
                        "type": "string"
                    },
                    "code": {
                        "description": "The code of the course",
                        "type": "string"
                    },
                    "total_hours": {
                        "description": "The total hours of the course",
                        "type": "string"
                    },
                    "published": {
                        "description": "The published status of the course",
                        "type": "boolean"
                    },
                    "free": {
                        "description": "The free status of the course",
                        "type": "boolean"
                    },
                    "price": {
                        "description": "The price of the course",
                        "type": "number",
                        "format": "float"
                    },
                    "available": {
                        "description": "The availability of the course",
                        "type": "boolean"
                    },
                    "image": {
                        "description": "The image of the course",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Course": {
                "title": "Course model",
                "description": "Course model",
                "required": [
                    "category_id",
                    "name"
                ],
                "properties": {
                    "category_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "total_hours": {
                        "type": "integer"
                    },
                    "free": {
                        "type": "boolean"
                    },
                    "published": {
                        "type": "boolean"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Course"
                }
            },
...
        }
    }
}